@@ -8,6 +8,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -34,7 +35,7 @@ public class BeerController {
     }
 
     @RequestMapping(method = RequestMethod.POST)
-    public ResponseEntity saveBeerObject(@RequestBody BeerDTO beer){
+    public ResponseEntity saveBeerObject(@Validated @RequestBody BeerDTO beer){
         BeerDTO savedBeer=beerService.saveBeerObject(beer);
 
         HttpHeaders headers = new HttpHeaders();
