@@ -1,9 +1,10 @@
 package com.systems.springframework.spring6restmvc.entities;
 
+import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.GenericGenerator;
 
-import javax.persistence.*;
+
 import java.time.LocalDateTime;
 import java.util.UUID;
 @Getter
